refactor(server): extract routes into domain-specific router files
Break up server.ts (~2900 lines) into 14 separate router files for better
maintainability and faster type inference:

- session.ts: Session CRUD, messages, diff, todo, summarize
- pty.ts: PTY management with WebSocket support
- provider.ts: Provider listing and OAuth flows
- mcp.ts: MCP server management and resources
- file.ts: File operations and search (find routes)
- config.ts: Configuration get/update
- tool.ts: Tool listing
- permission.ts: Permission requests
- auth.ts: Auth credentials
- worktree.ts: Worktree operations
- agent.ts: Agent listing
- lsp.ts: LSP status
- formatter.ts: Formatter status
- command.ts: Command listing

Expanded tui.ts to include all /tui/* routes with control routes nested.

This is a pure refactor - all 92 API endpoints preserved with identical
behavior. SDK regenerated with only ordering changes.
